t8_daq_system/requirements.txt
labjack-ljm>=1.23.0
matplotlib>=3.5.0
numpy>=1.21.0
